refactor: 권한 설정 완료, 이미 프로필이 존재하는 경우 예외 처리, 프로필 생성 응답 DTO 변경

+++ src/main/java/com/back/nbe12142team06/domain/user/repository/ClientProfileRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface ClientProfileRepository extends JpaRepository<ClientProfile, Integer> {
+public interface ClientProfileRepository extends JpaRepository<ClientProfile, Long> {
 
     @Modifying
     @Query("delete from ClientProfile p where p.user.id = :userId")
